requirements.txt
librosa
numpy
scipy
s3tokenizer
diffusers
torch==2.7.1
torchaudio==2.7.1
triton>=3.0.0
transformers==4.57.1
accelerate==1.10.1
onnxruntime
onnxruntime-gpu
einops
gradio

# Worker dependencies
boto3
requests
python-dotenv
soundfile
azure-storage-queue